pyproject.toml
[project]
name = "restate-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "restate-sdk",
    "hypercorn",
]
